Extract the mutation generator frim the MutationsCollectorVisitor into its dedicated class NodeMutationGenerator
